[WFLY-18370] Move the module tccl test into sub package
