+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/sar/context/classloader/module/ClassCInSarDeployment.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.test.integration.sar.context.classloader;
+package org.jboss.as.test.integration.sar.context.classloader.module;
 
 /**
  * @author: Jaikiran Pai
