+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/sar/context/classloader/module/jboss-deployment-structure.xml
