+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/sar/context/classloader/module/mbean/tccl-test-service.xml
@@ -23,6 +23,6 @@
         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:schemaLocation="urn:jboss:service:7.0 jboss-service_7_0.xsd">
 
-    <mbean name="wildfly:name=tccl-test-mbean" code="org.jboss.as.test.integration.sar.context.classloader.mbean.MBeanInAModuleService">
+    <mbean name="wildfly:name=tccl-test-mbean" code="org.jboss.as.test.integration.sar.context.classloader.module.mbean.MBeanInAModuleService">
     </mbean>
 </server>
